Added another tag to distinguish between what has been pushed to clearcase.
Before, calling gitcc checkin would move the clearcase tag which determines
where the next rebase is taken from. While this is simpler, it results in
the user having to be move careful about _always_ rebasing before a checkin.
It was also possible to double-up on changesets if the file was modified
more than once after the last rebase.

Now there are two tags - one for rebase and one for checkin.