Full rename support... finally!
Previously, as mentioned in the README, a 100% rename in Clearcase wasn't
supported in git-cc. This has been bugging me since the beginning. I finally got
around to tackling this, prompted by a big refactor at work which involved a
large number of straight renames across projects.

The solution isn't optimal, and I can't help thinking there must be a better
way to have implemented it. We know (and have always known) that a file has been
added to a directory. What we can't seem to find out is what version that file
is at a particular point in time. I would have thought/hoped that 'ls' might
have given us that information, but sadly no.

Instead it searches the history of the added file (which could be quite slow and
what I was trying to avoid doing) to find the last modified version before the
directories current timestamp. We skip this process if the file has also been
modified in the same changeset to save having to do the extra calculations.

Certainly far from perfect, but it's better than doing nothing.