Added --initial option to checkin.
When pushing to Clearcase from a git repository for the first time
there is no master_ci tag, which causes problems. To avoid this
the initial flag can be set to cause checkin to behave appropriately.

It would be possible to detect this, but I actually prefer this to be
a manual step.