Add symlink support to gitcc checkin.
This patches allows gitcc to create and modify symlinks within ClearCase during a checkin.

Note that git treats a change of symlink target as a modify. ClearCase does not allow modification of a symlink once it is created. This patch attempts to detect if a symlink already exists within the target ClearCase view, and will delete it first before adding the symlink. This works almost all of the time. One place it fails is when the symlink destination is outside of the ClearCase VOB. This causes problems, at least with ClearCase snapshot views. Although the symlink may be created, the user has to manually run 'cleartool update' on the view afterwards.

Signed-off-by: J. Tang <tang@jtang.org>