Fix "gitcc rebase" for empty files
(for files that had been checked in initially empty to CC)

Fixes #20.
The existing fix in charleso/git-cc@ccf628db033 is not working.

Solution:
If there is no "checkinversion" in "lshistory" then check if there is
3 items in the history (mkelemfile, mkelembranch, mkelemversion)
or 5 items (mkelemfile, mkelembranch, mkelemversion, mkbranchbranch, mkbranchversion)
If so then this file is created as an empty one and download it with proper version.