git use 'UTF-8' as default encoding
seeing : Git/doc/git/html/git-log.html

--encoding=<encoding>
The commit objects record the encoding used for the log message in their encoding header; this option can be used to tell the command to re-code the commit log message in the encoding preferred by the user. For non plumbing commands this defaults to UTF-8.